026.静态库和0警告

+++ Hazel/src/Hazel/ImGui/ImGuiLayer.cpp
@@ -65,7 +65,7 @@ namespace Hazel {
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2(static_cast<float>(app.GetWindow().GetWidth()), static_cast<float>(app.GetWindow().GetHeight()));
 		// Rendering
 		ImGui::Render();
 		ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
@@ -79,6 +79,9 @@ namespace Hazel {
 	}
 	void ImGuiLayer::OnImGuiRender()
 	{
+		//ImGui::Begin("Test");
+		//ImGui::Text("Hello World");
+		//ImGui::End();
 		static bool show = true;
 		ImGui::ShowDemoWindow(&show); // 当前OnImGuiRender层显示DemoUI窗口
 	}
